requirements.txt
pipecatcloud
pipecat-ai[cartesia,daily,openai,silero]>=0.0.69
python-dotenv~=1.0.1
